Added full threading model to the photon
With the addition of a timeout loop for publishing the particle cloud variables, now the photon will wait 5s for a connection - if not found, it will just keep on running the demo. Awesome. No more hanging until the photon is connected.

** But still, I get "no response from core" from the app even when the photon is reconnects after booting. I have moved the cloud variables to the loop() function so they would refresh at every cycle, but it doesn't seem to work.
Need to know a bit more about how to work with the cloud variables to make the cube completely 'self-updating'.
(Er, hoping maybe Kevin could shed some light on this subject for me.)